replace old chat and remove redundant conditional
