--- inc/research-guide-contact-new.php
@@ -1,56 +1,0 @@
-<div class="heading-holding-banner">
-    <a href="/contact/" title="Contact us">
-        <div class="heading-holding-banner">
-            <h2> <span> <span>Contact us for advice</span> </span> </h2>
-        </div>
-        <!-- end heading-holding-banner -->
-    </a>
-</div><!-- end .heading-holding-banner -->
-<div class="breather separator clr">
-    <h3 style="margin-bottom: 0;">Still need help?</h3>
-    <div class="breather clr">
-        <div class="width-full-to-three-quarters float-left">
-            <h3 class="margin-none margin-bottom-small"><a href="/contact/#livechat">Live chat</a></h3>
-            <p class="margin-none pad-none">For quick pointers<br/>Tuesday to Saturday<br />
-                09:00 to 17:00
-            </p>
-        </div>
-        <!-- end .width-full-to-three-quarters -->
-        <div class="width-full-to-one-quarter float-left">
-
-<div data-id="ePKP7MjW_D4" class="livechat_button"><a href="/contact-us/">live chat software</a></div>
-<!-- Start of LiveChat (www.livechatinc.com) code -->
-<script type="text/javascript">
-window.__lc = window.__lc || {};
-window.__lc.license = 10565762;
-(function() {
-  var lc = document.createElement('script'); lc.type = 'text/javascript'; lc.async = true;
-  lc.src = ('https:' == document.location.protocol ? 'https://' : 'http://') + 'cdn.livechatinc.com/tracking.js';
-  var s = document.getElementsByTagName('script')[0]; s.parentNode.insertBefore(lc, s);
-})();
-</script>
-<noscript>
-<a href="https://www.livechatinc.com/chat-with/10565762/" rel="nofollow">Chat with us</a>,
-powered by <a href="https://www.livechatinc.com/?welcome" rel="noopener nofollow" target="_blank">LiveChat</a>
-</noscript>
-<!-- End of LiveChat code -->
-
-        </div>
-        <!-- end .width-full-to-one-quarter -->
-    </div>
-    <!-- end .breather -->
-    <div class="breather clr">
-        <div class="width-full-to-three-quarters float-left">
-            <h3 class="margin-none margin-bottom-small"><a href="/contact-us/make-a-records-and-research-enquiry/" title="Email us">Email</a></h3>
-            <p class="margin-none pad-none">For more detailed research enquiries.</p>
-        </div>
-        <!-- end .width-full-to-three-quarters -->
-        <div class="width-full-to-one-quarter float-left">
-            <a href="/contact-us/make-a-records-and-research-enquiry/">
-                <img src="/wp-content/themes/tna/images/global/icon-128-email.png" alt="Email us">
-            </a>
-        </div>
-        <!-- end .width-full-to-one-quarter -->
-    </div>
-    <!-- end .breather -->
-</div><!-- end .breather -->
