+++ sibling-navigation.php
@@ -77,7 +77,7 @@ get_header(); ?>
     if (in_category("yes-contact")) {
     	?>
     	<?php
-        require_once('inc/research-guide-contact-new.php');
+        require_once('inc/research-guide-contact.php');
         ?>
         <?php
     }
